テストプロジェクトを一つに統一
