+++ cs-sdx.sln
@@ -9,8 +9,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sdx.DebugTool", "Sdx.DebugT
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTest", "UnitTest\UnitTest.csproj", "{9B61905B-00A2-43FE-ACD2-9571A358215B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSTest", "MSTest\MSTest.csproj", "{656B8D7F-B7EF-4DB4-BEDC-29EF79DD4BC7}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,10 +27,6 @@ Global
 		{9B61905B-00A2-43FE-ACD2-9571A358215B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9B61905B-00A2-43FE-ACD2-9571A358215B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9B61905B-00A2-43FE-ACD2-9571A358215B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{656B8D7F-B7EF-4DB4-BEDC-29EF79DD4BC7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{656B8D7F-B7EF-4DB4-BEDC-29EF79DD4BC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{656B8D7F-B7EF-4DB4-BEDC-29EF79DD4BC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{656B8D7F-B7EF-4DB4-BEDC-29EF79DD4BC7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
